requirements.txt
asn1crypto==0.24.0
attrs==19.3.0
backports-abc==0.5
backports.shutil-get-terminal-size==1.0.0
bleach==3.1.1
cachetools==3.1.1
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
configparser==4.0.2
contextlib2==0.6.0.post1
crayons==0.3.0
cryptography==2.1.4
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
enum34==1.1.9
ez-setup==0.9
functools32==3.2.3.post2
futures==3.3.0
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
gunicorn==19.10.0
httplib2==0.17.1
idna==2.9
importlib-metadata==1.5.0
ipaddress==1.0.23
ipykernel==4.10.1
ipython==5.9.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==5.2.0
jupyter-core==4.6.3
keyring==10.6.0
keyrings.alt==3.0
MarkupSafe==1.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==4.4.0
notebook==5.7.8
numpy==1.16.6
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.2.0.32
pandocfilters==1.4.2
pathlib2==2.3.5
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==1.0.18
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
Pygments==2.5.2
pygobject==3.26.1
pyrsistent==0.15.7
python-dateutil==2.8.1
pytube==9.6.0
pytz==2019.3
pyxdg==0.25
pyzmq==19.0.0
qtconsole==4.6.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scandir==1.10.0
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.14.0
terminado==0.8.3
testpath==0.4.4
tornado==5.1.1
traitlets==4.3.3
uritemplate==3.0.1
urllib3==1.25.8
wcwidth==0.1.8
webdriver-manager==2.3.0
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==1.2.0

